fix: shallowEqual, deepEqual 간에 반복되는 로직 util 함수로 분리
