+++ src/@lib/equalities/deepEquals.ts
@@ -1,26 +1,22 @@
 // deepEquals 함수는 두 값의 깊은 비교를 수행합니다.
 
+import { hasSameKey } from "./utils/hasSameKey";
+import { isValidObjects } from "./utils/isValidObjects";
+
 export function deepEquals<T>(objA: T, objB: T): boolean {
   if (Object.is(objA, objB)) {
     return true;
   }
 
-  if (
-    typeof objA !== "object" ||
-    objA === null ||
-    typeof objB !== "object" ||
-    objB === null
-  ) {
+  if (!isValidObjects(objA, objB)) {
+    return false;
+  }
+
+  if (!hasSameKey(objA, objB)) {
     return false;
   }
 
   const objAKeys = Object.keys(objA || {});
-  const objBKeys = Object.keys(objB || {});
-
-  if (objAKeys.length !== objBKeys.length) {
-    return false;
-  }
-
   for (const key of objAKeys) {
     // 재귀적으로 각 속성에 대해 deepEquals 호출
     if (!deepEquals(objA[key as keyof T], objB[key as keyof T])) {
