+++ src/@lib/equalities/shallowEquals.ts
@@ -1,6 +1,9 @@
 // shallowEquals 함수는 두 값의 얕은 비교를 수행합니다.
 // shallowEquals은 1뎁스까지만 비교한다.
 
+import { hasSameKey } from "./utils/hasSameKey";
+import { isValidObjects } from "./utils/isValidObjects";
+
 export function shallowEquals<T>(objA: T, objB: T): boolean {
   // 1. 두 값이 정확히 같은지 확인 (참조가 같은 경우)
   if (Object.is(objA, objB)) {
@@ -9,20 +12,14 @@ export function shallowEquals<T>(objA: T, objB: T): boolean {
 
   // 2. 둘 중 하나라도 객체가 아닌 경우 처리
   // - typeof null -> object 이기 떄문에 null check
-  if (
-    typeof objA !== "object" ||
-    objA === null ||
-    typeof objB !== "object" ||
-    objB === null
-  ) {
+  if (!isValidObjects(objA, objB)) {
     return false;
   }
 
   const objAKeys = Object.keys(objA || {});
-  const objBKeys = Object.keys(objB || {});
 
   // 3. 객체의 키 개수가 다른 경우 처리
-  if (objAKeys.length !== objBKeys.length) {
+  if (!hasSameKey(objA, objB)) {
     return false;
   }
