+++ src/@lib/equalities/utils/isValidObjects.ts
@@ -1,0 +1,12 @@
+export const isValidObjects = <T>(objA: T, objB: T) => {
+  if (
+    typeof objA !== "object" ||
+    objA === null ||
+    typeof objB !== "object" ||
+    objB === null
+  ) {
+    return false;
+  }
+
+  return true;
+};
